requirements.txt
fastapi==0.112.0
uvicorn==0.30.6
openai==1.40.0
pinecone==5.4.2
python-dotenv==1.0.1
pypdf==4.3.1
requests==2.32.3
tiktoken==0.7.0
